Контакты готовы.

+++ server/src/contacts/contacts.service.ts
@@ -1,15 +1,28 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { ContactEntity } from './contacts.entity';
+import { UsersService } from 'src/users/users.service';
 
 
 @Injectable()
 export class ContactsService {
-    constructor(private prisma: PrismaService) {}
+    constructor(private prisma: PrismaService, private usersService: UsersService) {}
 
 
   // Добавление контакта
-  async addContact(name: string, targetName: string, displayName: string): Promise<ContactEntity> {
+  async addContact(name: string, targetName: string, displayName: string | null) {
+    const userExists = await this.usersService.checkUserExistence(targetName);
+
+    if (!userExists) {
+      return false;
+    }
+
+    const contactExists = await this.contactExists(name, targetName);
+
+    if (contactExists) {
+      return false;
+    }
+
     const addedContact = await this.prisma.contact.create({
       data: {
         ownerName: name,
@@ -26,7 +39,7 @@ export class ContactsService {
 
 
   // Получение контактов пользователя
-  async getContacts(name: string): Promise<ContactEntity[]> {
+  async getContacts(name: string) {
     const foundContacts = await this.prisma.contact.findMany({
       where: {
         ownerName: name
@@ -40,8 +53,9 @@ export class ContactsService {
   }
 
 
-  // Проверка существования контакта
-  async contactExists(name: string, targetName: string): Promise<boolean> {
+  // Проверка существования контакта (Полезно, чтобы не дублировать новые записи о контакте при попытке
+  //                                  добавления уже существующего контакта)
+  async contactExists(name: string, targetName: string) {
     const foundContact = await this.prisma.contact.findUnique({
       where: {
         ownerName_targetName: {
@@ -56,7 +70,13 @@ export class ContactsService {
 
 
   // Обновление ОТОБРАЖАЕМОГО ИМЕНИ контакта
-  async updateContactDisplayName(name: string, targetName: string, displayName: string): Promise<boolean> {
+  async updateContactDisplayName(name: string, targetName: string, displayName: string) {
+    const userExists = await this.usersService.checkUserExistence(targetName);
+
+    if (!userExists) {
+      return false;
+    }
+
     const updatedContact = await this.prisma.contact.update({
       where: {
         ownerName_targetName: {
@@ -77,7 +97,13 @@ export class ContactsService {
 
 
   // Удаление контакта
-  async removeContact(name: string, targetName: string): Promise<boolean> {
+  async removeContact(name: string, targetName: string) {
+    const userExists = await this.usersService.checkUserExistence(targetName);
+
+    if (!userExists) {
+      return false;
+    }
+
     const removedContact = await this.prisma.contact.delete({
       where: {
         ownerName_targetName: {
